+++ lib/packages/docusaurus-shared/src/docusaurus-envhelper.ts
@@ -1,3 +1,5 @@
+import type { LoadContext, Plugin } from '@docusaurus/types';
+
 // Safe in browser & Node: check globalThis.process only by property access on a known object.
 // 1) Use Node's env if present; 2) else use browser-injected __DOCUSAURUS_ENV__; 3) else {}
 type Env = Record<string, string | undefined>;
@@ -50,4 +52,62 @@ export function addDocsRedir(redirectsArr: { to: string; from: string[] }[]) {
             from: [docUrl("/docs")],
         });
     }
+}
+
+
+
+
+
+interface HotjarThemeConfig {
+    applicationId: string;
+}
+
+export function pluginHotjar(context: LoadContext): Plugin {
+    const { siteConfig } = context;
+    const { themeConfig } = siteConfig;
+    const { hotjar } = themeConfig as { hotjar?: HotjarThemeConfig };
+
+    if (!hotjar) {
+        throw new Error(
+            `You need to specify 'hotjar' object in 'themeConfig' with 'applicationId' field in it to use docusaurus-plugin-hotjar`,
+        );
+    }
+
+    const { applicationId } = hotjar;
+
+    if (!applicationId) {
+        throw new Error(
+            'You specified the `hotjar` object in `themeConfig` but the `applicationId` field was missing. ' +
+            'Please ensure this is not a mistake.',
+        );
+    }
+
+    const isProd = process.env.NODE_ENV === 'production';
+
+    return {
+        name: 'docusaurus-plugin-hotjar',
+
+        injectHtmlTags() {
+            console.log(`[hotjar] applicationId = ${applicationId} isProd = ${isProd}`);
+            if (!isProd) {
+                return {};
+            }
+
+            return {
+                headTags: [
+                    {
+                        tagName: 'script',
+                        innerHTML: `(function(h,o,t,j,a,r){
+  h.hj=h.hj||function(){(h.hj.q=h.hj.q||[]).push(arguments)};
+  h._hjSettings={hjid:${applicationId},hjsv:6};
+  a=o.getElementsByTagName('head')[0];
+  r=o.createElement('script');r.async=1;
+  r.src=t+h._hjSettings.hjid+j+h._hjSettings.hjsv;
+  a.appendChild(r);
+})(window,document,'https://static.hotjar.com/c/hotjar-','.js?sv=');`,
+                    },
+                ],
+            };
+        },
+    };
 }
