+++ unified-doc/package.json
@@ -18,8 +18,9 @@
     "@docusaurus/core": "3.8.1",
     "@docusaurus/preset-classic": "3.8.1",
     "@docusaurus/theme-mermaid": "^3.8.1",
+    "@hotjar/browser": "^1.0.9",
     "@mdx-js/react": "^3.0.0",
-    "@openclint/docusaurus-shared": "^0.0.89",
+    "@openclint/docusaurus-shared": "^0.0.90",
     "asciinema-player": "^3.10.0",
     "clsx": "^2.0.0",
     "echarts": "^6.0.0",
